move Register instructions page comp into Register folder
